+++ src/main/java/uz/dev/rentcar/service/OfficeServiceImpl.java
@@ -2,13 +2,22 @@ package uz.dev.rentcar.service;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.CachePut;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.cache.annotation.Caching;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import uz.dev.rentcar.config.CaffeineCacheConfig;
+import uz.dev.rentcar.entity.Booking;
 import uz.dev.rentcar.entity.Office;
 import uz.dev.rentcar.entity.template.AbsLongEntity;
+import uz.dev.rentcar.exceptions.EntityNotDeleteException;
 import uz.dev.rentcar.mapper.OfficeMapper;
 import uz.dev.rentcar.payload.OfficeDTO;
 import uz.dev.rentcar.payload.response.PageableDTO;
@@ -16,23 +25,30 @@ import uz.dev.rentcar.repository.OfficeRepository;
 import uz.dev.rentcar.service.template.OfficeService;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class OfficeServiceImpl implements OfficeService {
 
     private final OfficeMapper officeMapper;
+
     private final OfficeRepository officeRepository;
 
     @Override
+    @Cacheable(value = CaffeineCacheConfig.OFFICES, key = "#id")
     public OfficeDTO read(Long id) {
 
         Office office = officeRepository.getByIdOrThrow(id);
+
+        log.info("read office {}", office);
 
         return officeMapper.toDTO(office);
     }
 
     @Override
+    @Cacheable(value = CaffeineCacheConfig.OFFICES, key = "#page + '-' + #size")
     public PageableDTO readAll(int page, int size) {
 
         Sort sort = Sort.by(AbsLongEntity.Fields.id).ascending();
@@ -42,6 +58,8 @@ public class OfficeServiceImpl implements OfficeService {
         Page<Office> officePage = officeRepository.findAll(pageable);
 
         List<Office> offices = officePage.getContent();
+
+        log.info("read all offices {}", offices);
 
         return new PageableDTO(
                 officePage.getSize(),
@@ -55,17 +73,31 @@ public class OfficeServiceImpl implements OfficeService {
 
     @Override
     @Transactional
+    @Caching(
+            put = {@CachePut(value = CaffeineCacheConfig.OFFICES, key = "#result.id")},
+            evict = {
+                    @CacheEvict(value = CaffeineCacheConfig.OFFICES, allEntries = true)
+            }
+    )
     public OfficeDTO create(OfficeDTO officeDTO) {
 
         Office office = officeMapper.toEntity(officeDTO);
 
         officeRepository.save(office);
 
+        log.info("create office {}", office);
+
         return officeMapper.toDTO(office);
     }
 
     @Override
     @Transactional
+    @Caching(
+            put = {@CachePut(value = CaffeineCacheConfig.OFFICES, key = "#id")},
+            evict = {
+                    @CacheEvict(value = CaffeineCacheConfig.OFFICES, allEntries = true)
+            }
+    )
     public OfficeDTO update(Long id, OfficeDTO officeDTO) {
 
         Office office = officeRepository.getByIdOrThrow(id);
@@ -77,14 +109,32 @@ public class OfficeServiceImpl implements OfficeService {
 
         officeRepository.save(office);
 
+        log.info("update office {}", office);
+
         return officeMapper.toDTO(office);
     }
 
     @Override
     @Transactional
+    @Caching(evict = {
+            @CacheEvict(value = CaffeineCacheConfig.OFFICES, allEntries = true)
+    })
     public void delete(Long id) {
 
         Office office = officeRepository.getByIdOrThrow(id);
+
+        List<Booking> pickupBookings = office.getPickupBookings();
+
+        List<Booking> returnBookings = office.getReturnBookings();
+
+        if ((Objects.nonNull(pickupBookings) && !pickupBookings.isEmpty()) ||
+                (Objects.nonNull(returnBookings) && !returnBookings.isEmpty())) {
+
+            throw new EntityNotDeleteException("Cannot delete office with existing pickup bookings", HttpStatus.BAD_REQUEST);
+
+        }
+
+        log.info("delete office {}", office);
 
         officeRepository.delete(office);
     }
